requirements.txt
streamlit
paho-mqtt
mysql.connector
pandas